+++ scripts/dotnet/Dll.cs
@@ -1,0 +1,17 @@
+/// Copyright (c)  2023  Xiaomi Corporation (authors: Fangjun Kuang)
+/// Copyright (c)  2023 by manyeyes
+/// Copyright (c)  2024.5 by 东风破
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System;
+
+namespace SherpaOnnx
+{
+    internal static class Dll
+    {
+        public const string Filename = "sherpa-onnx-c-api";
+    }
+}
