+++ scripts/dotnet/OfflineTransducerModelConfig.cs
@@ -1,0 +1,30 @@
+/// Copyright (c)  2024.5 by 东风破
+
+using System.Linq;
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
+using System.Text;
+using System;
+
+namespace SherpaOnnx
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct OfflineTransducerModelConfig
+    {
+        public OfflineTransducerModelConfig()
+        {
+            Encoder = "";
+            Decoder = "";
+            Joiner = "";
+        }
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Encoder;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Decoder;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Joiner;
+    }
+
+}
