+++ scripts/dotnet/OfflineWhisperModelConfig.cs
@@ -1,0 +1,37 @@
+/// Copyright (c)  2024.5 by 东风破
+
+using System.Linq;
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
+using System.Text;
+using System;
+
+namespace SherpaOnnx
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct OfflineWhisperModelConfig
+    {
+        public OfflineWhisperModelConfig()
+        {
+            Encoder = "";
+            Decoder = "";
+            Language = "";
+            Task = "transcribe";
+            TailPaddings = -1;
+        }
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Encoder;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Decoder;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Language;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Task;
+
+        public int TailPaddings;
+    }
+
+}
